refactor: Update answer prompt in quizRoute.js
This commit updates the answer prompt in the quizRoute.js file to remove the specific context of being a CTO of a company hiring fresher software developers. The prompt now simply instructs the user to rate the answers on a scale of 1-10 and provide the response as an array of ratings. This change improves the clarity and flexibility of the prompt.